refactor: moved out d3 methods to own directory
